requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
sentence_transformers
langchain_huggingface 
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
torch
arxiv 
wikipedia
duckduckgo-search
google-search-results
validators
youtube_transcript_api
pytube
unstructured
numexpr
huggingface-hub